resolved code review comments

+++ src/main/java/com/smarthost/reservation/serviceinterface/IReserveService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import com.smarthost.reservation.dto.RoomUsageDetailsDTO;
 
 public interface IReserveService {
-	public void initGuestPriceQuotes(List<Integer> guestPriceQuotes);
+	public void initGuestPriceQuotes(List<Float> guestPriceQuotes);
 
 	public RoomUsageDetailsDTO getRoomUsageDetails(int inputPremiumRoomCount, int inputEconomyRoomCount);
